fix: KV cache pointer offset overflow in rmsnorm_rope and flash_decode_split kernels
Cast cache_token_idx and physical_block_id to int64 before multiplying
with stride values, preventing signed 32-bit integer overflow when
KV cache offset exceeds 2^31 (triggered at per-rank batch_size >= 8).

+++ python/triton_dist/mega_triton_kernel/kernels/flash_decode.py
@@ -100,7 +100,7 @@ def attn_gqa_fwd_batch_decode_split_kv_task_para(
         current_token_indices = logical_block_idx * PAGE_SIZE + offs_page
         mask_valid_tokens = current_token_indices < cur_kv_seq_len
 
-        physical_block_id = tl.load(block_table_ptr + bid * stride_table_bs + logical_block_idx)
+        physical_block_id = tl.load(block_table_ptr + bid * stride_table_bs + logical_block_idx).to(tl.int64)  # 避免大batch下32位地址偏移溢出
         kv_off = physical_block_id * (PAGE_SIZE * stride_cache_bs) + kv_hid * stride_cache_h + offs_in_page
         k_tile = tl.load(k_cache_ptr + kv_off)
         v_tile = tl.load(v_cache_ptr + kv_off)
